requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-multipart==0.0.9
jinja2==3.1.3
cryptography==42.0.2
pydantic==2.6.1
python-dotenv==1.0.1 